// frontend/src/types/skin.ts
export type Skin = {
  id: string; // skinId
  name: string; // AWP | Dragon Lore
  rarity: string; // Covert
  collection: string; // The ... Collection
  wear: string; // Factory New
  float: number; // 0.05231
  price: number; // $100.34
  isStatTrak: boolean;
  wasWon: boolean;
  imgSrc: string;
  createdAt: Date;
}